+++ rust/tests/integration/pipe_transport.rs
@@ -18,6 +18,7 @@ use mediasoup::rtp_parameters::{
 };
 use mediasoup::sctp_parameters::SctpStreamParameters;
 use mediasoup::srtp_parameters::{SrtpCryptoSuite, SrtpParameters};
+use mediasoup::transport::ProduceError;
 use mediasoup::webrtc_transport::{TransportListenIps, WebRtcTransport, WebRtcTransportOptions};
 use mediasoup::worker::{RequestError, Worker, WorkerSettings};
 use mediasoup::worker_manager::WorkerManager;
@@ -549,6 +550,39 @@ fn pipe_to_router_succeeds_with_video() {
             ],
         );
         assert!(pipe_producer.paused());
+    });
+}
+
+#[test]
+fn pipe_to_router_fails_if_both_routers_belong_to_the_same_worker() {
+    future::block_on(async move {
+        let (worker1, _worker2, router1, _router2, transport1, _transport2) = init().await;
+
+        let router1bis = worker1
+            .create_router(RouterOptions::new(media_codecs()))
+            .await
+            .expect("Failed to create router");
+
+        let video_producer = transport1
+            .produce(video_producer_options())
+            .await
+            .expect("Failed to produce video");
+
+        let result = router1
+            .pipe_producer_to_router(
+                video_producer.id(),
+                PipeToRouterOptions::new(router1bis.clone()),
+            )
+            .await;
+
+        if let Err(PipeProducerToRouterError::ProduceFailed(ProduceError::Request(
+            RequestError::Response { reason },
+        ))) = result
+        {
+            assert!(reason.contains("already exists [method:transport.produce]"));
+        } else {
+            panic!("Unexpected result: {:?}", result);
+        }
     });
 }