+++ worker/src/RTC/DataConsumer.cpp
@@ -2,12 +2,10 @@
 // #define MS_LOG_DEV_LEVEL 3
 
 #include "RTC/DataConsumer.hpp"
-#include "ChannelMessageHandlers.hpp"
 #include "DepLibUV.hpp"
 #include "Logger.hpp"
 #include "MediaSoupErrors.hpp"
 #include "Utils.hpp"
-#include "Channel/ChannelNotifier.hpp"
 #include "RTC/SctpAssociation.hpp"
 #include <stdexcept>
 
@@ -16,14 +14,15 @@ namespace RTC
 	/* Instance methods. */
 
 	DataConsumer::DataConsumer(
+	  RTC::Shared* shared,
 	  const std::string& id,
 	  const std::string& dataProducerId,
 	  RTC::SctpAssociation* sctpAssociation,
 	  RTC::DataConsumer::Listener* listener,
 	  json& data,
 	  size_t maxMessageSize)
-	  : id(id), dataProducerId(dataProducerId), sctpAssociation(sctpAssociation), listener(listener),
-	    maxMessageSize(maxMessageSize)
+	  : id(id), dataProducerId(dataProducerId), shared(shared), sctpAssociation(sctpAssociation),
+	    listener(listener), maxMessageSize(maxMessageSize)
 	{
 		MS_TRACE();
 
@@ -67,7 +66,7 @@ namespace RTC
 			this->protocol = jsonProtocolIt->get<std::string>();
 
 		// NOTE: This may throw.
-		ChannelMessageHandlers::RegisterHandler(
+		this->shared->channelMessageRegistrator->RegisterHandler(
 		  this->id,
 		  /*channelRequestHandler*/ this,
 		  /*payloadChannelRequestHandler*/ this,
@@ -78,7 +77,7 @@ namespace RTC
 	{
 		MS_TRACE();
 
-		ChannelMessageHandlers::UnregisterHandler(this->id);
+		this->shared->channelMessageRegistrator->UnregisterHandler(this->id);
 	}
 
 	void DataConsumer::FillJson(json& jsonObject) const
@@ -209,10 +208,10 @@ namespace RTC
 				// Trigger 'bufferedamountlow' now.
 				if (this->bufferedAmount <= this->bufferedAmountLowThreshold)
 				{
-					std::string data(R"({"bufferedAmount":")");
+					std::string data(R"({"bufferedAmount":)");
 
 					data.append(std::to_string(this->bufferedAmount));
-					data.append("\"}");
+					data.append("}");
 				}
 				// Force the trigger of 'bufferedamountlow' once there is less or same
 				// buffered data than the given threshold.
@@ -350,13 +349,20 @@ namespace RTC
 			this->forceTriggerBufferedAmountLow = false;
 
 			// Notify the Node DataConsumer.
-			std::string data(R"({"bufferedAmount":")");
+			std::string data(R"({"bufferedAmount":)");
 
 			data.append(std::to_string(this->bufferedAmount));
-			data.append("\"}");
+			data.append("}");
 
-			Channel::ChannelNotifier::Emit(this->id, "bufferedamountlow", data);
+			this->shared->channelNotifier->Emit(this->id, "bufferedamountlow", data);
 		}
+	}
+
+	void DataConsumer::SctpAssociationSendBufferFull()
+	{
+		MS_TRACE();
+
+		this->shared->channelNotifier->Emit(this->id, "sctpsendbufferfull");
 	}
 
 	// The caller (Router) is supposed to proceed with the deletion of this DataConsumer
@@ -369,7 +375,7 @@ namespace RTC
 
 		MS_DEBUG_DEV("DataProducer closed [dataConsumerId:%s]", this->id.c_str());
 
-		Channel::ChannelNotifier::Emit(this->id, "dataproducerclose");
+		this->shared->channelNotifier->Emit(this->id, "dataproducerclose");
 
 		this->listener->OnDataConsumerDataProducerClosed(this);
 	}