+++ rust/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+# 0.11.2
+
+* Updates from mediasoup TypeScript `3.10.11..=3.11.2`.
+
+# 0.11.1
+
+* Updates from mediasoup TypeScript `3.10.7..=3.10.10`.
+
 # 0.11.0
 
 * Updates from mediasoup TypeScript `3.10.2..=3.10.6`.